backend/requirements.txt
asgiref==3.7.2
Django==4.2.16
django-cors-headers==4.3.1
djangorestframework==3.15.2
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
Pillow==10.4.0
